Add method to retrieve from URLs concurrently
retrieveContentAtURLs uses a few of the concurrency features of java 8 to
perform concurrent retrieval of the content at the supplied URLs.

We create a task to execute for each URL and leverage an execution service to
run them for us. We use the CompletableFuture allOf method to block until
completion, and then aggregate the results.

Signed-off-by: Moses Mendoza <moses@puppet.com>